:recycle: feature/abstractive_loginPage_and_BestCombiantionDetailPage
