+++ src/pages/BestCombinationDetailPage.tsx
@@ -6,21 +6,6 @@ import { 인터페이스_꿀조합 } from '@typings/ISandwich';
 import { collection, getDoc, doc } from 'firebase/firestore';
 import { db } from '../firebase.config';
 
-const 꿀조합_데이터_가져오기 = async (꿀조합id: string) => {
-  try {
-    const 꿀조합_콜랙션 = collection(db, '꿀조합');
-    const querySnapshot = await getDoc(doc(꿀조합_콜랙션, 꿀조합id));
-    if (querySnapshot.exists()) {
-      const 꿀조합 = querySnapshot.data();
-      return 꿀조합;
-    }
-    console.log('꿀조합이 없습니다.!');
-  } catch (error) {
-    console.error(error);
-    console.log('꿀조합 가져오기 실패');
-  }
-};
-
 function BestCombinationDetailPage() {
   // TODO: any 없애기
   const 꿀조합: 인터페이스_꿀조합 | any = useLoaderData();
@@ -48,5 +33,20 @@ function BestCombinationDetailPage() {
 }
 
 export default BestCombinationDetailPage;
+
+const 꿀조합_데이터_가져오기 = async (꿀조합id: string) => {
+  try {
+    const 꿀조합_콜랙션 = collection(db, '꿀조합');
+    const querySnapshot = await getDoc(doc(꿀조합_콜랙션, 꿀조합id));
+    if (querySnapshot.exists()) {
+      const 꿀조합 = querySnapshot.data();
+      return 꿀조합;
+    }
+    console.log('꿀조합이 없습니다.!');
+  } catch (error) {
+    console.error(error);
+    console.log('꿀조합 가져오기 실패');
+  }
+};
 
 export const loader = ({ params }: LoaderFunctionArgs) => 꿀조합_데이터_가져오기(params.combinationId!);
